requirements.txt
Flask
openai==0.28
python-docx
PyMuPDF
mammoth
langchain
python-dotenv
gunicorn
